added line to whatsnew v0.19.2 and test to test_missing.py in series folder
